+++ packages/video_player/video_player/test/video_player_test.dart
@@ -4,12 +4,13 @@
 
 import 'dart:async';
 import 'dart:io';
+
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:video_player/video_player.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:video_player/video_player.dart';
 import 'package:video_player_platform_interface/video_player_platform_interface.dart';
 
 class FakeController extends ValueNotifier<VideoPlayerValue>
@@ -26,23 +27,31 @@ class FakeController extends ValueNotifier<VideoPlayerValue>
 
   @override
   String get dataSource => '';
+
   @override
   DataSourceType get dataSourceType => DataSourceType.file;
+
   @override
   String get package => null;
+
   @override
   Future<Duration> get position async => value.position;
 
   @override
   Future<void> seekTo(Duration moment) async {}
+
   @override
   Future<void> setVolume(double volume) async {}
+
   @override
   Future<void> initialize() async {}
+
   @override
   Future<void> pause() async {}
+
   @override
   Future<void> play() async {}
+
   @override
   Future<void> setLooping(bool looping) async {}
 
@@ -51,6 +60,9 @@ class FakeController extends ValueNotifier<VideoPlayerValue>
 
   @override
   Future<ClosedCaptionFile> get closedCaptionFile => _loadClosedCaption();
+
+  @override
+  bool get useCache => false;
 }
 
 Future<ClosedCaptionFile> _loadClosedCaption() async =>
@@ -176,32 +188,71 @@ void main() {
             });
       });
 
-      test('network', () async {
-        final VideoPlayerController controller = VideoPlayerController.network(
-          'https://127.0.0.1',
-        );
-        await controller.initialize();
-
-        expect(
-            fakeVideoPlayerPlatform.dataSourceDescriptions[0],
-            <String, dynamic>{
-              'uri': 'https://127.0.0.1',
-              'formatHint': null,
-            });
-      });
-
-      test('network with hint', () async {
-        final VideoPlayerController controller = VideoPlayerController.network(
+      group('network', () {
+        test('with cache', () async {
+          final VideoPlayerController controller =
+              VideoPlayerController.network(
             'https://127.0.0.1',
-            formatHint: VideoFormat.dash);
-        await controller.initialize();
+            useCache: true,
+          );
+          await controller.initialize();
 
-        expect(
-            fakeVideoPlayerPlatform.dataSourceDescriptions[0],
-            <String, dynamic>{
-              'uri': 'https://127.0.0.1',
-              'formatHint': 'dash',
-            });
+          expect(
+              fakeVideoPlayerPlatform.dataSourceDescriptions[0],
+              <String, dynamic>{
+                'uri': 'https://127.0.0.1',
+                'formatHint': null,
+                'useCache': true,
+              });
+        });
+
+        test('without cache', () async {
+          final VideoPlayerController controller =
+              VideoPlayerController.network(
+            'https://127.0.0.1',
+            useCache: false,
+          );
+          await controller.initialize();
+
+          expect(
+              fakeVideoPlayerPlatform.dataSourceDescriptions[0],
+              <String, dynamic>{
+                'uri': 'https://127.0.0.1',
+                'formatHint': null,
+                'useCache': false,
+              });
+        });
+
+        test('without cache by default', () async {
+          final VideoPlayerController controller =
+              VideoPlayerController.network(
+            'https://127.0.0.1',
+          );
+          await controller.initialize();
+
+          expect(
+              fakeVideoPlayerPlatform.dataSourceDescriptions[0],
+              <String, dynamic>{
+                'uri': 'https://127.0.0.1',
+                'formatHint': null,
+                'useCache': false,
+              });
+        });
+
+        test('network with hint', () async {
+          final VideoPlayerController controller =
+              VideoPlayerController.network('https://127.0.0.1',
+                  formatHint: VideoFormat.dash);
+          await controller.initialize();
+
+          expect(
+              fakeVideoPlayerPlatform.dataSourceDescriptions[0],
+              <String, dynamic>{
+                'uri': 'https://127.0.0.1',
+                'formatHint': 'dash',
+                'useCache': false,
+              });
+        });
       });
 
       test('init errors', () async {