+++ packages/video_player/video_player_platform_interface/lib/video_player_platform_interface.dart
@@ -56,7 +56,7 @@ abstract class VideoPlayerPlatform {
   ///
   /// This method is called when the plugin is first initialized
   /// and on every full restart.
-  Future<void> init() {
+  Future<void> init(int maxCacheSize, int maxCacheFileSize) {
     throw UnimplementedError('init() has not been implemented.');
   }
 
@@ -135,13 +135,16 @@ class DataSource {
   ///
   /// The [package] argument must be non-null when the asset comes from a
   /// package and null otherwise.
+  ///
+  /// The [useCache] argument must be non-null, default is false.
   DataSource({
     @required this.sourceType,
     this.uri,
     this.formatHint,
     this.asset,
     this.package,
-  });
+    this.useCache = false,
+  }) : assert(useCache != null);
 
   /// The way in which the video was originally loaded.
   ///
@@ -165,6 +168,9 @@ class DataSource {
   /// The package that the asset was loaded from. Only set for
   /// [DataSourceType.asset] videos.
   final String package;
+
+  /// Use cache for this data source or not. Used only for network data source.
+  final bool useCache;
 }
 
 /// The way in which the video was originally loaded.