+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.6
+
+* Add caching functionality to videos from network sources.
+
 ## 1.0.5
 
 * Make the pedantic dev_dependency explicit.